Refactor quote merge observer/handler into the sales namespace
